==> notebooks/0.EDA_data.ipynb <==
import pandas as pd
import os
import re
import xml.etree.ElementTree as ET

kaggle = pd.read_csv('../data/data_remaning_kaggle.csv', sep = ',', index_col = 'index')
kaggle = kaggle.iloc[:,1:]

kaggle.to_csv('../data/kaggle.csv')

exported = pd.read_csv('../data/exported_db_data.csv', sep = ',', index_col = 'index')

tree = ET.parse('../data/data_xml_clean.xml')
root = tree.getroot()

xml_list = []
for child in root:
    xml_dict = {}
    for subchild in child:
        xml_dict[subchild.tag] = subchild.text
    xml_list.append(xml_dict)

xml_series = pd.Series(xml_list) # creamos serie

df_xml = pd.DataFrame.from_dict(xml_series)  # creamos DF malo a partir de los diccionarios

para_limpiar_xml = df_xml[0].apply(pd.Series)  # se convierten k-v en filas-columnas

para_limpiar_xml = para_limpiar_xml.drop(columns =["level_0"]) 

para_limpiar_xml.set_index("index", inplace=True)  # cambiamos el indice

para_limpiar_xml.to_csv('../data/df_xml.csv')  # guardamos en csv

texto = pd.read_csv("../data/data_txt_clean.txt", sep=";", index_col = 'index')  

def preguntas(archivo):
    print(f"¿Cuántas filas y columnas tenemos en cada uno de los dataframes?\n{archivo.shape}")
    print(f"¿Cuáles son los tipos de datos de cada una de las columnas?\n")
    display(archivo.info())
    print(f"¿Cuáles son los elementos básicos estadísticos de los dataframes?\n")
    display(archivo.describe())
    print(f"¿Qué información tenemos?\n")
    display(archivo.head())
    

preguntas(kaggle)

preguntas(xml)

preguntas(exported)

preguntas(texto)
