==> notebooks/1. fichero_unificado.ipynb <==
import pandas as pd
import numpy as np

xml = pd.read_csv('../data/df_xml.csv', index_col = 'index')
txt = pd.read_csv("../data/data_txt_clean.txt", sep = ";", index_col = 'index')
kaggle = pd.read_csv('../data/kaggle.csv', index_col='index')
exported = pd.read_csv('../data/exported_db_data.csv', index_col='index')

raw_data = xml.join(txt).join(exported).join(kaggle)
raw_data.head(2)

raw_data.to_csv("../data/raw_data.csv")
