==> notebooks/2. df_analisis.ipynb <==
Vamos con XML

import pandas as pd
import numpy as np
import re
import sidetable

columnas_select = ['time', 'age', 'gender', 'Q3', 'Q4', 'Q5', 'Q6', 'Q7' 'Q20', 'Q25',
       'Q39_Part_1', 'Q39_Part_2', 'Q39_Part_3', 'Q39_Part_4', 'Q39_Part_5',
       'Q39_Part_6', 'Q39_Part_7', 'Q39_Part_8', 'Q39_Part_9', 'Q39_OTHER',
       'Q40_Part_1', 'Q40_Part_2', 'Q40_Part_3', 'Q40_Part_4', 'Q40_Part_5',
       'Q40_Part_6', 'Q40_Part_7', 'Q40_Part_8', 'Q40_Part_9', 'Q40_Part_10',
       'Q40_Part_11', 'Q40_OTHER', 'Q42_Part_1', 'Q42_Part_2', 'Q42_Part_3',
       'Q42_Part_4', 'Q42_Part_5', 'Q42_Part_6', 'Q42_Part_7', 'Q42_Part_8',
       'Q42_Part_9', 'Q42_Part_10', 'Q42_Part_11', 'Q42_OTHER']

df_uni = pd.read_csv('../data/raw_data.csv', index_col = 0)
pd.options.display.max_columns = None
df_uni.head(1)

df_adalab = df_uni[['time', 'age', 'gender', 'Q3', 'Q4', 'Q5', 'Q6', 'Q7', 'Q9',
       'Q10_Part_1', 'Q10_Part_2', 'Q10_Part_3', 'Q10_Part_4', 'Q10_Part_5', 'Q10_Part_6', 'Q10_Part_7', 'Q10_Part_8', 'Q10_Part_9', 'Q10_Part_10',
       'Q10_Part_11', 'Q10_Part_12', 'Q10_Part_13', 'Q10_Part_14', 'Q10_Part_15', 'Q10_Part_16', 'Q10_OTHER',
       'Q14', 'Q20', 'Q25',
       'Q32_B_Part_1', 'Q32_B_Part_2', 'Q32_B_Part_3', 'Q32_B_Part_4', 'Q32_B_Part_5',
       'Q32_B_Part_6', 'Q32_B_Part_7', 'Q32_B_Part_8', 'Q32_B_Part_9', 'Q32_B_Part_10',
       'Q32_B_Part_11', 'Q32_B_Part_12', 'Q32_B_Part_13', 'Q32_B_Part_14', 'Q32_B_Part_15',
       'Q32_B_Part_16', 'Q32_B_Part_17', 'Q32_B_Part_18', 'Q32_B_Part_19', 'Q32_B_Part_20', 'Q32_B_OTHER',
       'Q33', 'Q35',
       'Q39_Part_1', 'Q39_Part_2', 'Q39_Part_3', 'Q39_Part_4', 'Q39_Part_5',
       'Q39_Part_6', 'Q39_Part_7', 'Q39_Part_8', 'Q39_Part_9', 'Q39_OTHER',
       'Q40_Part_1', 'Q40_Part_2', 'Q40_Part_3', 'Q40_Part_4', 'Q40_Part_5',
       'Q40_Part_6', 'Q40_Part_7', 'Q40_Part_8', 'Q40_Part_9', 'Q40_Part_10',
       'Q40_Part_11', 'Q40_OTHER', 'Q42_Part_1', 'Q42_Part_2', 'Q42_Part_3',
       'Q42_Part_4', 'Q42_Part_5', 'Q42_Part_6', 'Q42_Part_7', 'Q42_Part_8',
       'Q42_Part_9', 'Q42_Part_10', 'Q42_Part_11', 'Q42_OTHER']]

df_adalab.head(1)

df_adalab.columns

#Vamos a cambiar el nombre de nuestras columnas

columnas_antiguas = ['time', 'age', 'gender', 'Q3', 'Q4', 'Q5', 'Q6', 'Q7', 'Q9',
       'Q10_Part_1', 'Q10_Part_2', 'Q10_Part_3', 'Q10_Part_4', 'Q10_Part_5',
       'Q10_Part_6', 'Q10_Part_7', 'Q10_Part_8', 'Q10_Part_9', 'Q10_Part_10',
       'Q10_Part_11', 'Q10_Part_12', 'Q10_Part_13', 'Q10_Part_14',
       'Q10_Part_15', 'Q10_Part_16', 'Q10_OTHER',
       'Q14', 'Q20', 'Q25',
       'Q32_B_Part_1', 'Q32_B_Part_2', 'Q32_B_Part_3', 'Q32_B_Part_4',
       'Q32_B_Part_5', 'Q32_B_Part_6', 'Q32_B_Part_7', 'Q32_B_Part_8',
       'Q32_B_Part_9', 'Q32_B_Part_10', 'Q32_B_Part_11', 'Q32_B_Part_12',
       'Q32_B_Part_13', 'Q32_B_Part_14', 'Q32_B_Part_15', 'Q32_B_Part_16',
       'Q32_B_Part_17', 'Q32_B_Part_18', 'Q32_B_Part_19', 'Q32_B_Part_20', 'Q32_B_OTHER',
       'Q33', 'Q35',
       'Q39_Part_1', 'Q39_Part_2', 'Q39_Part_3', 'Q39_Part_4', 'Q39_Part_5',
       'Q39_Part_6', 'Q39_Part_7', 'Q39_Part_8', 'Q39_Part_9', 'Q39_OTHER',
       'Q40_Part_1', 'Q40_Part_2', 'Q40_Part_3', 'Q40_Part_4', 'Q40_Part_5',
       'Q40_Part_6', 'Q40_Part_7', 'Q40_Part_8', 'Q40_Part_9', 'Q40_Part_10', 'Q40_Part_11', 'Q40_OTHER', 'Q42_Part_1',
       'Q42_Part_2', 'Q42_Part_3', 'Q42_Part_4', 'Q42_Part_5', 'Q42_Part_6',
       'Q42_Part_7', 'Q42_Part_8', 'Q42_Part_9', 'Q42_Part_10', 'Q42_Part_11',
       'Q42_OTHER']

nuevas_columnas = ['time', 'age', 'gender', 'country', 'education', 'role', 'coding_experience', 'prog_language', 'code_env', 
'notebook_kaggle', 'notebook_colab', 'notebook_azure', 'notebook_paperspace', 'notebook_binder','notebook_codeocean', 'notebook_ibmwatsonstudio',
'notebook_amazonsagemaker', 'notebook_amazonemr', 'notebook_googlecloudnotebooks','notebook_googleclouddatalab', 'notebook_databricks', 'notebook_zeppelin',
'notebook_deepnote','notebook_observable', 'notebook_none', 'notebook_OTHER',
'viz', 'field', 'salary',
'learn_dp_MySQL', 'learn_dp_PostgreSQL', 'learn_dp_SQLite', 'learn_dp_Oracle', 'learn_dp_Mongo', 'learn_dp_Snowflake', 'learn_dp_IBMDd2', 'learn_dp_SQLServer',
'learn_dp_AzureSQL', 'learn_dp_AzureCosmos', 'learn_dp_Redshift', 'learn_dp_Aurora',
'learn_dp_RDS', 'learn_dp_DynamoDB', 'learn_dp_BigQuery', 'learn_dp_CloudSQL', 'learn_dp_Firestore', 'learn_dp_BigTable', 'learn_dp_Spanner', 'learn_dp_none', 'learn_dp_Other',
'dp_most_freq', 'BI_tools',
'share_plotlydash', 'share_streamlit', 'share_NBViewer', 'share_GitHub', 'share_Personal blog', 'share_Kaggle', 'share_Colab', 'share_Shiny', 'share_None', 'share_OTHER',
'courses_Coursera', 'courses_edX', 'courses_Kaggle', 'courses_DataCamp', 'courses_Fastai', 'courses_Udacity', 'courses_Udemy', 'courses_LinkedIn',
'courses_Cloud-certification', 'courses_University', 'courses_None', 'courses_OTHER',
'media_Twitter', 'media_Email', 'media_Reddit', 'media_Kaggle', 'media_Forums',
'media_YouTube', 'media_Podcasts', 'media_Blogs', 'media_Journal', 'media_Slack', 'media_None', 'media_OTHER']

columnas_ok = dict(zip(columnas_antiguas, nuevas_columnas))

df_adalab.rename(columns = columnas_ok, inplace = True)

df_adalab.head(1)

# cambiamos el nombre de uno de los valores para evitar confusiones
# Jupyter (JupyterLab, Jupyter Notebooks, etc)

df_adalab["code_env"] = df_adalab["code_env"].str.replace("Jupyter (JupyterLab, Jupyter Notebooks, etc)", "JupyterLab")

df_adalab["code_env"].value_counts()

#miramos los valores nulos de nuestros datos
df_adalab.stb.missing()

# Viendo que hay columnas con 100% nulos, vamos a eliminarlas porque no aportan valor:

df_adalab.drop(['media_None', 'learn_dp_none', 'courses_None', 'notebook_none'], axis= 1, inplace= True)

df_adalab.stb.missing()

#miramos cuantas filas tenemos en total
df_adalab.shape

#miramos cuantas profesiones tenemos y cuántos registros de cada profesion
df_adalab['role'].value_counts()

#miramos cuales son los rangos de coding_experience. Vamos a reorganizar estas categorías y a darles nombres más comprensibles.
df_adalab['coding_experience'].unique()

def experience(experiencia):
    if experiencia == 'I have never written code':
        return 'Sin experiencia'
    elif experiencia == '< 1 years' or experiencia == '1-3 years':
        return 'Junior'
    elif experiencia == '3-5 years':
        return 'Mid'
    elif  experiencia == '5-10 years':
        return 'Senior'
    elif experiencia == '10-20 years' or experiencia == '20+ years':
        return 'Senior+'

df_adalab['cat_experience'] = df_adalab['coding_experience'].apply(experience)
df_adalab.head(2)

# vamos a investigar la columna time puesto que no nos cuadra mucho el dato:

df_adalab["time"].value_counts()

#Eliminamos las columnas iniciales.

df_adalab.drop(['coding_experience', "time"], axis = 1, inplace = True)

df_adalab.to_csv("../data/df_adalab_limpio.csv")
