==> notebooks/3. creacion_df_filtrados.ipynb <==
LÍNEA DE INVESTIGACIÓN: Marketing
PREGUNTAS A RESPONDER:
- Género: Q2
- Edad de las mujeres de la muestra -> Q1
- En qué países europeos hay mayor número de mujeres en el mundo del data -> Q3
- Nivel de estudios de las mujeres de la muestra -> Q4
- Sector del que vienen las mujeres de la muestra
- Rol -> Q5
- Sector de la empresa actual -> Q20
- Salario -> Q25
- Plataformas de cursos utilizadas -> Q40
- Redes más utilizadas -> Q42
- Plataformas uso de compartir "código" -> Q39

### PREGUNTAS

PASO UNO (CÓMO LLEGO A MI PÚBLICO OBJETIVO)
- ¿Hay poco porcentaje de mujeres en el sector tecnológico? ¿Esto está relacionado con el país que analizamos (global, europero, español)? **Resuelto con gráfica
- Salario de Data analist y Bussines Analyst (en función de la experiencia). Brecha de género.
- Analizar el nivel de estudios y el sector del que vienen las mujeres que trabajan en DA o BA.

PASO DOS (QUÉ LE OFREZCO A MI PÚBLICO OBJETIVO)

- seleccionar rol: Data Analist, Business Analist

- con qué trabajan esos roles: herramientas de visualizacion, de graficas, notebooks, bases de datos relacionales y no relacionales, lenguajes de programación.

import pandas as pd
import numpy as np
import os
import seaborn as sns
import matplotlib.pyplot as plt
import sidetable

df_adalab = pd.read_csv('../data/df_adalab_limpio.csv', index_col = 'index')
pd.options.display.max_columns = None
df_adalab.head(2)

# Vamos a crear nuevo dataframe filtrado por los roles que nos interesan: da/ba:
df_adalab['role'].unique()

df_da = df_adalab[(df_adalab['role'] == 'Data Analyst') | (df_adalab['role'] == 'Business Analyst')]
df_da.head(2)

#Vamos a ver los rangos salariales que tenemos para DA/BA:
df_da['salary'].value_counts()

#miramos cuales son los rangos de salary. Vamos a reorganizar estas categorías para no tener tantas y darles mayor sentido (salarios comunes en funcion de la experiencia).
df_da['salary'].unique()

def salario(sal):
    try:
        if sal in ['$0-999', '1,000-1,999', '2,000-2,999', '3,000-3,999', '4,000-4,999', '5,000-7,499', '7,500-9,999', '10,000-14,999', '15,000-19,999']:
            return '< 20K'
        elif sal in ['20,000-24,999','25,000-29,999', '30,000-39,999', '40,000-49,999']:
            return '20K - 50K'
        elif sal in ['50,000-59,999', '60,000-69,999', '70,000-79,999', '80,000-89,999', '90,000-99,999']:
            return '50K - 100K'
        elif sal in ['100,000-124,999', '125,000-149,999', '150,000-199,999']:
            return '100K - 200K'
        elif sal in ['200,000-249,999','250,000-299,999', '300,000-499,999','$500,000-999,999' , '>$1,000,000']:
            return '> 200K'
    except:
        return np.nan

df_da['cat_salary'] = df_da['salary'].apply(salario)
df_da.head(2)

df_da[df_da["country"] == "Spain"]["country"].count()

df_adalab.drop(["salary"], axis = 1, inplace = True)

df_da.to_csv('../data/df_da.csv')

european_latin_countries = ["Albania", "Andorra", "Austria", "Belarus", "Belgium", "Bosnia and Herzegovina", "Bulgaria", "Croatia", "Cyprus", "Czech Republic", "Denmark", "Estonia", "Finland", "France", "Germany", "Greece", "Hungary", "Iceland", "Ireland", "Italy", "Latvia", "Liechtenstein", "Lithuania", "Luxembourg", "Malta", "Moldova", "Monaco", "Montenegro", "Netherlands", "North Macedonia", "Norway", "Poland", "Portugal", "Romania", "Russia", "San Marino", "Serbia", "Slovakia", "Slovenia", "Spain", "Sweden", "Switzerland", "Ukraine", "United Kingdom", "Vatican City", 
                            "Argentina", "Bolivia", "Brazil", "Chile", "Colombia", "Costa Rica", "Cuba", "Dominican Republic", "Ecuador", "El Salvador", "Guatemala", "Honduras", "Mexico", "Nicaragua", "Panama", "Paraguay", "Peru", "Uruguay", "Venezuela"]

europa = ["Albania", "Andorra", "Austria", "Belarus", "Belgium", "Bosnia and Herzegovina", "Bulgaria", "Croatia", "Cyprus", "Czech Republic", "Denmark", "Estonia", "Finland", "France", "Germany", "Greece", "Hungary", "Iceland", "Ireland", "Italy", "Latvia", "Liechtenstein", "Lithuania", "Luxembourg", "Malta", "Moldova", "Monaco", "Montenegro", "Netherlands", "North Macedonia", "Norway", "Poland", "Portugal", "Romania", "Russia", "San Marino", "Serbia", "Slovakia", "Slovenia", "Spain", "Sweden", "Switzerland", "Ukraine", "United Kingdom", "Vatican City"]

df_da_europa = df_da[df_da["country"].isin (europa)]

df_da_europa.to_csv('../data/df_da_europa.csv')

df_publicidad = df_adalab[((df_adalab['role'] == 'Other') | (df_adalab['role'] == 'Currently not employed') | (df_adalab['role'] == 'Student')) & (df_adalab['country'].isin(european_latin_countries))]

df_publicidad.head(2)

df_da_publi = df_adalab[((df_adalab['role'] == 'Data Analyst') | (df_adalab['role'] == 'Business Analyst')) & (df_adalab['country'].isin(european_latin_countries))]

df_publicidad.to_csv('../data/df_publicidad.csv')

df_da_publi.to_csv('../data/df_da_publi.csv')
