==> notebooks/4. situacion_actual.ipynb <==
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np

import sidetable
from sklearn.experimental import enable_iterative_imputer
from sklearn.impute import IterativeImputer
from sklearn.impute import KNNImputer

df_da = pd.read_csv('../data/df_da.csv', index_col = 'index')
pd.options.display.max_columns = None
df_da.sample(4)

df_da.describe(include='object').T

df_da.stb.missing()

# creamos diferentes df con datos sociodemográficos propios para cada columna que nos interesa 
# y poder expandirlas sin crear demasiadas filas


sociodemo_language = df_da[["age","gender", "country", "education", "role", "field", "cat_experience", "cat_salary","prog_language"] ]
sociodemo_language["languages"] = sociodemo_language["prog_language"].str.split(",")
df_languages = sociodemo_language.explode(["languages"])
df_languages.drop("prog_language", inplace= True, axis= 1)
df_languages.sample(2)

sociodemo_code_env = df_da[["age","gender", "country", "education", "role", "field", "cat_experience", "cat_salary","code_env"] ]
sociodemo_code_env["cod_env"] = sociodemo_code_env["code_env"].str.split(",")
df_code_env = sociodemo_code_env.explode(["cod_env"])
df_code_env.drop("code_env", inplace= True, axis= 1)
df_code_env.sample(2)

sociodemo_viz = df_da[["age","gender", "country", "education", "role", "field", "cat_experience", "cat_salary","viz"] ]
sociodemo_viz["visual"] = sociodemo_viz["viz"].str.split(",")
df_viz = sociodemo_viz.explode(["visual"])
df_viz.drop("viz", inplace= True, axis= 1)
df_viz.sample(2)

#Creamos la función para crear una columna que una las anteriores:
def unir(col):
    lista_valores = []
    for element in col:
        if str(element) != "nan":
            lista_valores.append(str(element))
        else:
            pass
    return lista_valores

# creamos una lista con todas las columnas relacionadas a la pregunta notebooks

lista_notebooks = ['notebook_kaggle', 'notebook_colab', 'notebook_azure',
       'notebook_paperspace', 'notebook_binder', 'notebook_codeocean',
       'notebook_ibmwatsonstudio', 'notebook_amazonsagemaker',
       'notebook_amazonemr', 'notebook_googlecloudnotebooks',
       'notebook_googleclouddatalab', 'notebook_databricks',
       'notebook_zeppelin', 'notebook_deepnote', 'notebook_observable',
       'notebook_OTHER']

# aplicamos la función de unir columnas sobre el df completo pero uniendo solo las columnas 'notebooks':
df_da['notebooks_unido'] = df_da[lista_notebooks].apply(unir, axis=1)

#comprobamos:
df_da.head(2)

# Nos vamos a quedar solo con las columnas sociodemograficas y la nueva creada:

sociodemo_notebooks = df_da[["age","gender", "country", "education", "role", "field", "cat_experience", "cat_salary","notebooks_unido"]]
df_notebooks = sociodemo_notebooks.explode(["notebooks_unido"])
df_notebooks.head()

# vamos a comprobar si quedan nulos en esta columna (muestra que no haya usado nunca uno de estos notebooks):

df_notebooks['notebooks_unido'].isnull().sum()

# vamos a eliminar estas filas (solo queremos saber los notebooks más utilizados):

df_notebooks.dropna(subset= 'notebooks_unido', axis = 0, inplace= True)

df_notebooks['notebooks_unido'].isnull().sum()

# vamos a hacer lo mismo con las columnas learn:

col_lear_dp = ['learn_dp_MySQL', 'learn_dp_PostgreSQL', 'learn_dp_SQLite', 'learn_dp_Oracle', 'learn_dp_Mongo', 'learn_dp_Snowflake', 'learn_dp_IBMDd2', 'learn_dp_SQLServer',
'learn_dp_AzureSQL', 'learn_dp_AzureCosmos', 'learn_dp_Redshift', 'learn_dp_Aurora',
'learn_dp_RDS', 'learn_dp_DynamoDB', 'learn_dp_BigQuery', 'learn_dp_CloudSQL', 'learn_dp_Firestore', 'learn_dp_BigTable', 'learn_dp_Spanner', 'learn_dp_Other']

df_da['learn_unido'] = df_da[col_lear_dp].apply(unir, axis=1)

sociodemo_learn = df_da[["age","gender", "country", "education", "role", "field", "cat_experience", "cat_salary","learn_unido"]]
df_learn = sociodemo_learn.explode(["learn_unido"])
df_learn.head()

df_learn['learn_unido'].isnull().sum()

df_learn.dropna(subset= 'learn_unido', axis = 0, inplace= True)

df_learn['learn_unido'].isnull().sum()

col_share = ['share_plotlydash', 'share_streamlit', 'share_NBViewer', 'share_GitHub', 'share_Personal blog', 'share_Kaggle', 'share_Colab', 'share_Shiny', 'share_None', 'share_OTHER']

df_da['share_unido'] = df_da[col_share].apply(unir, axis=1)

sociodemo_share = df_da[["age","gender", "country", "education", "role", "field", "cat_experience", "cat_salary","share_unido"]]
df_share = sociodemo_share.explode(["share_unido"])
df_share.head()

df_share['share_unido'].isnull().sum()

df_share.dropna(subset= 'share_unido', axis = 0, inplace= True)

df_share['share_unido'].isnull().sum()

col_courses = ['courses_Coursera', 'courses_edX', 'courses_Kaggle', 'courses_DataCamp', 'courses_Fastai', 'courses_Udacity', 'courses_Udemy', 'courses_LinkedIn',
'courses_Cloud-certification', 'courses_University', 'courses_OTHER']

df_da['courses_unido'] = df_da[col_courses].apply(unir, axis=1)

sociodemo_courses = df_da[["age","gender", "country", "education", "role", "field", "cat_experience", "cat_salary","courses_unido"]]
df_courses = sociodemo_courses.explode(["courses_unido"])
df_courses.head()

df_courses['courses_unido'].isnull().sum()

df_courses.dropna(subset= 'courses_unido', axis = 0, inplace= True)

df_courses['courses_unido'].isnull().sum()

col_media = ['media_Twitter', 'media_Email', 'media_Reddit', 'media_Kaggle', 'media_Forums', 'media_YouTube', 'media_Podcasts', 'media_Blogs', 'media_Journal', 'media_Slack', 'media_OTHER']

df_da['media_unido'] = df_da[col_media].apply(unir, axis=1)

sociodemo_media = df_da[["age","gender", "country", "education", "role", "field", "cat_experience", "cat_salary","media_unido"]]
df_media = sociodemo_media.explode(["media_unido"])
df_media.head()

df_media['media_unido'].isnull().sum()

df_media.dropna(subset= 'media_unido', axis = 0, inplace= True)

df_media['media_unido'].isnull().sum()

df_languages.to_csv("../data/df_languages.csv")
df_viz.to_csv("../data/df_viz.csv")
df_code_env.to_csv("../data/df_code_env.csv")
df_notebooks.to_csv("../data/df_notebooks.csv")
df_learn.to_csv("../data/df_learn.csv")
df_share.to_csv("../data/df_share.csv")
df_courses.to_csv("../data/df_courses.csv")
df_media.to_csv("../data/df_media.csv")
